requirements.txt
streamlit
google-generativeai
Pillow
gTTS
SpeechRecognition